% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer.slimr.R
\name{get_lmer_convergence}
\alias{get_lmer_convergence}
\title{returns TRUE if model converged without warnings and FALSE otherwise}
\usage{
get_lmer_convergence(m)
}
\description{
returns TRUE if model converged without warnings and FALSE otherwise
}
