% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer.util.R
\name{lmer.check.convergence}
\alias{lmer.check.convergence}
\title{Returns a (g)lmer model if it converges, and a warning or error otherwise}
\usage{
lmer.check.convergence(formula, data, family = "gaussian",
  use.old.lme4 = F, ...)
}
\description{
Returns a (g)lmer model if it converges, and a warning or error otherwise
}
