\name{lmer.slimr}
\alias{lmer.slimr}
\title{Evaluate sequence of lmer models with a decreasing number of random
correlations and return the first one to converge.}
\usage{
  lmer.slimr(formula, data, parallel = 1,
    show.warnings = F, ...)
}
\description{
  Evaluate sequence of lmer models with a decreasing number
  of random correlations and return the first one to
  converge.
}

