\name{ct}
\alias{ct}
\title{for continuous variables. should be called in the model formula, to ensure
that variable is centered on the relevant data subset}
\usage{
  ct(x)
}
\description{
  for continuous variables. should be called in the model
  formula, to ensure that variable is centered on the
  relevant data subset
}

