\name{glmer.slimr}
\alias{glmer.slimr}
\title{Evaluate sequence of glmer models with a decreasing number of random
correlations and return the first one to converge.}
\usage{
  glmer.slimr(formula, data, family, parallel = 1,
    show.warnings = F, ...)
}
\description{
  Evaluate sequence of glmer models with a decreasing
  number of random correlations and return the first one to
  converge.
}

