% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lmer.check.convergence}
\alias{lmer.check.convergence}
\title{Returns a (g)lmer model if it converges, and a warning or error otherwise}
\usage{
lmer.check.convergence(formula, data, family = "gaussian", use.old.lme4 = F,
  ...)
}
\description{
Returns a (g)lmer model if it converges, and a warning or error otherwise
}

