\name{lmer.check.convergence}
\alias{lmer.check.convergence}
\title{Returns a (g)lmer model if it converges, and a warning or error otherwise}
\usage{
  lmer.check.convergence(formula, data,
    family = "gaussian", ...)
}
\description{
  Returns a (g)lmer model if it converges, and a warning or
  error otherwise
}

