% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{lmer.slimr}
\alias{lmer.slimr}
\alias{lmer.slimr-package}
\title{Evaluate sequence of lmer models with a decreasing number of random
correlations and return the first one to converge.}
\usage{
lmer.slimr(formula, data, parallel = getOption("mc.cores", 2L),
  suppress.warnings = F, simplest.only = F, return.only.converged = F,
  use.old.lme4 = F, ...)
}
\description{
Evaluate sequence of lmer models with a decreasing number of random
correlations and return the first one to converge.

lmer.slimr: iteratively slim down (g)lmer models
}

