\name{lmer.slimr}
\alias{lmer.slimr}
\title{Evaluate sequence of (g)lmer models with a decreasing number of random
correlations and return the first one to converge.}
\usage{
  lmer.slimr(formula, parallel, ...)
}
\description{
  Evaluate sequence of (g)lmer models with a decreasing
  number of random correlations and return the first one to
  converge.
}

